[cmake] Converted to 'modern' CMake
As a results, most CMakeLists file are mcuh shorter.

This includes:

- use of imported targets for the dependencies (allows to clean up the
CMakeLists - no need to write into the 'include_directories' for
instance)
- OpenFace now exports its own targets (Utilities, LandmarkDetector, FaceAnalyser, GazeAnalyser)
- generation and installation of OpenFaceConfig.cmake & co (this make it possible to
properly include OpenFace in other projects)

Based on https://rix0r.nl/blog/2015/08/13/cmake-guide/

Additional changes:
- on non-Windows platforms, compile shared libraries
- libraries now have proper versioning
- if dlib is installed in the system, use it. Otherwise compile it.

Issues:

- if dlib is not available, cmake fails saying that 'dlib' is not in
the export set of OpenFace. This can be easily fixed by adding an
install target to the dlib CMakeLists (but since it is a 3rd party lib,
I do not want to modify the source). Alternatively, updating the version
of dlib to a more recent version (>18.18, I think) should fix it, as
dlib has then added a proper install target to its CMakeLists.

- tested on Ubuntu 17.10 only!